+++ app/src/main/java/com/example/y_trackcomercial/usecases/marcacionPromotora/getIdVisitaActivaUseCase.kt
@@ -7,7 +7,7 @@ class GetIdVisitaActivaUseCase @Inject constructor(
     private val visitasRepository: VisitasRepository
 ) {
     // suspend
-    suspend  fun getActiveVisitId(): Int {
+    suspend  fun getActiveVisitId(): Long {
         return visitasRepository.getIdVisitaActiva()
     }
 }
